refactor: JumpingPiece 추상화
