+++ src/main/java/domain/piece/King.java
@@ -1,13 +1,11 @@
 package domain.piece;
 
 import domain.Direction;
-import domain.Square;
 import domain.Team;
 
 import java.util.List;
-import java.util.Map;
 
-public class King extends Piece {
+public class King extends JumpingPiece {
     public King(final Team team) {
         super(team);
     }
@@ -23,15 +21,5 @@ public class King extends Piece {
                 Direction.NORTH_WEST,
                 Direction.SOUTH_EAST,
                 Direction.SOUTH_WEST);
-    }
-
-    @Override
-    public boolean canMove(final Square source, final Square target, final Map<Square, Piece> pieces) {
-        for (final Direction movableDirection : movableDirections()) {
-            if (source.next(movableDirection).equals(target)) {
-                return true;
-            }
-        }
-        return false;
     }
 }
