+++ src/main/java/domain/piece/Knight.java
@@ -1,13 +1,11 @@
 package domain.piece;
 
 import domain.Direction;
-import domain.Square;
 import domain.Team;
 
 import java.util.List;
-import java.util.Map;
 
-public class Knight extends Piece {
+public class Knight extends JumpingPiece {
     public Knight(final Team team) {
         super(team);
     }
@@ -24,15 +22,5 @@ public class Knight extends Piece {
                 Direction.WEST_WEST_NORTH,
                 Direction.WEST_WEST_SOUTH
         );
-    }
-
-    @Override
-    public boolean canMove(final Square source, final Square target, final Map<Square, Piece> pieces) {
-        for (final Direction movableDirection : movableDirections()) {
-            if (source.next(movableDirection).equals(target)) {
-                return true;
-            }
-        }
-        return false;
     }
 }
